requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
alembic==1.13.2
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==23.2.0
Babel==2.15.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
behave==1.2.6
bleach==6.1.0
blinker==1.7.0
Brotli==1.1.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
croniter==3.0.3
cryptography==43.0.0
cssselect==1.2.0
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
execnet==2.1.1
fasteners==0.19
filelock==3.15.4
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Mailman==1.1.1
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-Security-Too==5.5.2
flask-sqlacodegen==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flask_simplelogin==0.1.3
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
inflect==7.3.1
iniconfig==2.0.0
ipinfo==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
kaitaistruct==0.10
lxml==5.2.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkdocs-material-extensions==1.3.1
more-itertools==10.3.0
multidict==6.0.5
numpy==2.0.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
parse==1.20.2
parse-type==0.6.2
passlib==1.7.4
pdbp==1.5.3
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pony==0.7.17
psutil==6.0.0
psycopg2-binary==2.9.9
py==1.11.0
pyasn1==0.6.0
pycparser==2.22
pydivert==2.1.0
Pygments==2.18.0
pynose==1.5.2
pyOpenSSL==24.2.1
pyotp==2.9.0
pyparsing==3.1.2
pyreadline3==3.4.1
PySocks==1.7.1
pytest==8.3.2
pytest-html==2.0.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==14.0
pytest-xdist==3.6.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
sbvirtualdisplay==1.3.0
selenium==4.23.1
selenium-wire==5.1.0
seleniumbase==4.29.4
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
speaklater==1.3
SQLAlchemy==2.0.32
SQLAlchemy-Utils==0.41.2
tabcompleter==1.3.3
trio==0.26.0
trio-websocket==0.11.1
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
undetected-chromedriver==3.5.5
urllib3==2.2.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
wsproto==1.2.0
WTForms==3.1.2
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.19.2
zstandard==0.23.0
fake_useragent